requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
chromadb
tiktoken
python-dotenv